+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Log of files not copied or moved.
+
 ### Changed
-- Refactored into a class. 
+- Refactored into a class.
+
+### Fixed
+- `move_file` now checks if source exists, destination exists and if so that the destination is an actual file
+  before calling `diff_files`.
 
 ## [0.6.0] - 2020-02-10
 ### Added